+++ BetterCrafting/Models/Theme.cs
@@ -24,4 +24,16 @@ public class Theme {
 
 	[JsonConverter(typeof(ColorConverter))]
 	public Color? QuantityWarningShadowColor { get; set; }
+
+
+	public bool CustomTooltip { get; set; }
+
+	public bool CustomScroll { get; set; }
+
+	[JsonConverter(typeof(ColorConverter))]
+	public Color? TextColor { get; set; }
+
+	[JsonConverter(typeof(ColorConverter))]
+	public Color? TextShadowColor { get; set; }
+
 }